+++ components/date.tsx
@@ -1,6 +1,6 @@
 import { parseISO, format } from "date-fns";
 
-export default function Date({ dateString }) {
+export default function Date({ dateString }: { dateString: string }) {
   const date = parseISO(dateString);
   return <time dateTime={dateString}>{format(date, "LLLL d, yyyy")}</time>;
 }
