+++ pages/api/hello.ts
@@ -1,0 +1,5 @@
+import { NextApiRequest, NextApiResponse } from "next";
+
+export default (_: NextApiRequest, res: NextApiResponse) => {
+  res.status(200).json({ text: "Hello" });
+};
