--- pages/posts/first-post.js
@@ -1,19 +1,0 @@
-import Head from "next/head";
-import Link from "next/link";
-import Layout from "../../components/layout";
-
-export default function FirstPost() {
-  return (
-    <Layout>
-      <Head>
-        <title>First Post</title>
-      </Head>
-      <h1>First Post</h1>
-      <h2>
-        <Link href="/">
-          <a>Back to home</a>
-        </Link>
-      </h2>
-    </Layout>
-  );
-}
